Frep first year problem fix
The problem with historic Frep is that stock synthesis does not provide
the next year quarter needed to calculate the final year. Adding the
Frep calc to the final year re-run should have worked but delivered zero
for a result. However, this is due to a bug in the model initialisation
where EByQtrLastYr was not initialised, hence zero and hence the Frep
value was always zero for the final assessment year.